+++ scripts/generate_import_cache_cpp.py
@@ -182,7 +182,7 @@ for name, value in json_data.items():
 
 for file in files:
     content = file.to_string()
-    path = f'src/include/duckdb_python/import_cache/modules/{file.file_name}'
+    path = f'src/duckdb_py/include/duckdb_python/import_cache/modules/{file.file_name}'
     import_cache_path = os.path.join(script_dir, '..', path)
     with open(import_cache_path, "w") as f:
         f.write(content)
@@ -237,7 +237,9 @@ private:
 
 """
 
-import_cache_path = os.path.join(script_dir, '..', 'src/include/duckdb_python/import_cache/python_import_cache.hpp')
+import_cache_path = os.path.join(
+    script_dir, '..', 'src/duckdb_py/include/duckdb_python/import_cache/python_import_cache.hpp'
+)
 with open(import_cache_path, "w") as f:
     f.write(import_cache_file)
 
@@ -252,7 +254,7 @@ def get_module_file_path_includes(files: List[ModuleFile]):
 module_includes = get_module_file_path_includes(files)
 
 modules_header = os.path.join(
-    script_dir, '..', 'src/include/duckdb_python/import_cache/python_import_cache_modules.hpp'
+    script_dir, '..', 'src/duckdb_py/include/duckdb_python/import_cache/python_import_cache_modules.hpp'
 )
 with open(modules_header, "w") as f:
     f.write(module_includes)
