+++ tests/fast/relational_api/test_rapi_description.py
@@ -10,7 +10,7 @@ class TestRAPIDescription(object):
         types = [x[1] for x in desc]
         assert names == ['a', 'b']
         assert types == ['INTEGER', 'BIGINT']
-        assert (all([x == duckdb.NUMBER for x in types]))
+        assert all([x == duckdb.NUMBER for x in types])
 
     def test_rapi_describe(self, duckdb_cursor):
         np = pytest.importorskip("numpy")
